Fix DeleteSkill functionality

+++ src/main/java/command/DeleteSkill.java
@@ -1,6 +1,7 @@
 package command;
 
 import services.SkillService;
+import utils.CheckSkills;
 import view.View;
 
 import java.sql.SQLException;
@@ -9,6 +10,7 @@ public class DeleteSkill implements Command {
     public static final String DELETE_SKILL = "delete skill";
     private final View view;
     SkillService skillService = new SkillService();
+    CheckSkills checkSkills = new CheckSkills();
 
     public DeleteSkill(View view) {
         this.view = view;
@@ -22,18 +24,22 @@ public class DeleteSkill implements Command {
     @Override
     public void execute() {
         int id;
-        while (true) {
-            try {
-                view.write("Please, enter skill id to delete: ");
-                id = Integer.parseInt(view.read());
-                break;
-            } catch (NumberFormatException e) {
-                view.write("Invalid value. Use digits");
-            }
-        }
         try {
-            skillService.deleteSkill(id);
-            view.write("Skill with id " + id + " successfully deleted");
+            while (true) {
+                try {
+                    view.write("Please, enter skill id to delete: ");
+                    id = Integer.parseInt(view.read());
+                    if (checkSkills.IsSkillIdExists(id)) {
+                        skillService.deleteSkill(id);
+                        view.write("Skill with id " + id + " successfully deleted");
+                        break;
+                    } else {
+                        System.out.println("Skill id doesn't exists");
+                    }
+                } catch (NumberFormatException e) {
+                    view.write("Invalid value. Use digits");
+                }
+            }
         } catch (SQLException e) {
             e.getStackTrace();
         }
